% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmdpref_1.R
\name{cmdpref}
\alias{cmdpref}
\title{MDPREF model}
\usage{
cmdpref(pref, ndim = 2, monotone = FALSE, tor = 1e-08, maxit = 50)
}
\arguments{
\item{pref}{Preference data matrix. Rows as objects and columns as subjects. Small value for less preferred and large value for more preferred.}

\item{ndim}{Number of dimensions.}

\item{monotone}{TRUE for Kruskal monotonic transformation. FALSE for metric scale.}

\item{tor}{tolerance for monotonic transformation.}

\item{maxit}{maximum number of interactions for monotonic transformation.}
}
\value{
An object of class "mdpref".
\item{score }{Object coordinates}
\item{corr }{Subject vectors}
\item{d }{Singular values}
\item{fitted }{fitted preference matrix}
\item{tpref }{transformed preference matrix}
\item{stress }{Stress I value under monotonic transformation}
}
\description{
Metric and non-metric MDPREF model
}
\details{
Metric scale:\cr
Singular decomposition is applied to data matrix S.\cr
S = ULV', X = U*sqrt(n-1),
Y = V*sqrt(L)/sqrt(n-1), n = no. of objects.\cr \cr
Non-metric scale:\cr
Kruskal monotonic transformation is applied to each subject vector by opscale(level=2,...) in optiscale package.\cr
Metric MDPREF is applied to the transformed data.
}
\examples{
library(lattice)
library(optiscale)
library(cmdpref)
#rank order of preferences of 5 objects for 4 subjects
mydata

fit<-cmdpref(pref=mydata,monotone=TRUE,maxit=100)
summary(fit)

#biplot() arguments are applicable
plot(fit,xlim=c(-2,2),cex=1)
}
\references{
Carroll, J. D. (1972). “Individual Differences and Multidimensional Scaling.” In Multidimensional Scaling: Theory and Applications in the Behavioral Sciences, vol. 1, edited by R. N. Shepard, A. K. Romney, and S. B. Nerlove, 105–155. New York: Seminar Press.\cr\cr
Kruskal, Joseph B. (1964) “Nonmetric Multidimensional Scaling: A Numerical Method.” Psychometrika 29: 115-129
}
\author{
Chi-wai Kwan
}
