% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmdpref_1.R
\name{plot.mdpref}
\alias{plot.mdpref}
\title{biplot for mdperf}
\usage{
\method{plot}{mdpref}(x, ...)
}
\arguments{
\item{x}{mdpref object}

\item{...}{biplot() parameters}
}
\description{
biplot for mdperf
}
