% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmdpref_1.R
\name{summary.mdpref}
\alias{summary.mdpref}
\title{summary for mdpref}
\usage{
\method{summary}{mdpref}(object, ...)
}
\arguments{
\item{object}{mdpref object}

\item{...}{summary parameters}
}
\description{
summary for mdpref
}
